NEXA-Agentic-AI/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.98.0
motor==3.7.1
pymongo==4.14.0
python-dotenv==1.1.1
